Move sqlite unit tests to the net/tests directory
